+++ src/UI/ItineraryControlUI.js
@@ -1,5 +1,5 @@
 /*
-Copyright - 2017 2021 - wwwouaiebe - Contact: https://www.ouaie.be/
+Copyright - 2017 2022 - wwwouaiebe - Contact: https://www.ouaie.be/
 
 This  program is free software;
 you can redistribute it and/or modify it under the terms of the
@@ -166,7 +166,7 @@ class ItineraryControlUI {
 	@type {ManeuversCheckboxInputEL}
 	*/
 
-	#maneuversCheckboxInputEL = null
+	#maneuversCheckboxInputEL = null;
 
 	/**
 	A reference to the HTMLElement in witch the control have to be added