+++ src/UI/OsmSearchWaitUI.js
@@ -1,5 +1,5 @@
 /*
-Copyright - 2017 2021 - wwwouaiebe - Contact: https://www.ouaie.be/
+Copyright - 2017 2022 - wwwouaiebe - Contact: https://www.ouaie.be/
 
 This  program is free software;
 you can redistribute it and/or modify it under the terms of the
@@ -44,23 +44,16 @@ class OsmSearchWaitUI {
 	#waitDiv;
 
 	/**
-	The red bullet...
-	@type {HTMLElement}
-	*/
-
-	#waitBullet;
-
-	/**
 	The constructor
 	*/
 
 	constructor ( ) {
 		this.#waitDiv = theHTMLElementsFactory.create (
 			'div',
-			{ className : 'TravelNotes-WaitAnimation' },
+			{ className : 'TravelNotes-WaitAnimation' }
 		);
 
-		this.#waitBullet = theHTMLElementsFactory.create (
+		theHTMLElementsFactory.create (
 			'div',
 			{
 				className : 'TravelNotes-WaitAnimationBullet'