+++ src/data/Maneuver.js
@@ -1,5 +1,5 @@
 /*
-Copyright - 2017 2021 - wwwouaiebe - Contact: https://www.ouaie.be/
+Copyright - 2017 2022 - wwwouaiebe - Contact: https://www.ouaie.be/
 
 This  program is free software;
 you can redistribute it and/or modify it under the terms of the
@@ -96,7 +96,7 @@ class Maneuver extends TravelObject {
 	@type {Number}
 	*/
 
-	#objId = INVALID_OBJ_ID;;
+	#objId = INVALID_OBJ_ID;
 
 	/**
 	the constructor