+++ src/data/TravelNotesData.js
@@ -1,5 +1,5 @@
 /*
-Copyright - 2017 2021 - wwwouaiebe - Contact: https://www.ouaie.be/
+Copyright - 2017 2022 - wwwouaiebe - Contact: https://www.ouaie.be/
 
 This  program is free software;
 you can redistribute it and/or modify it under the terms of the
@@ -56,7 +56,7 @@ class Routing {
 	@type {String}
 	*/
 
-	#transitMode = ''
+	#transitMode = '';
 
 	/**
 	The constructor